les10

+++ src/main/java/ru/geekbrains/summer/market/dto/ProductDto.java
@@ -3,13 +3,16 @@ package ru.geekbrains.summer.market.dto;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import ru.geekbrains.summer.market.model.Product;
+
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 public class ProductDto {
     private Long id;
     private String title;
     private String categoryTitle;
-    private int price;
+    private BigDecimal price;
 
     public ProductDto(Product product) {
         this.id = product.getId();
